backend/requirements.txt
# Core framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# Database
sqlalchemy>=2.0.0
psycopg[binary]>=3.1.0
alembic>=1.12.0
pgvector>=0.2.0

# Validation and settings
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.0.0

# Environment
python-dotenv>=1.0.0

# AI/ML
openai>=1.3.0

# HTTP client
httpx>=0.25.0

# Auth
passlib[bcrypt]>=1.7.4
PyJWT>=2.8.0# File upload (multipart/form-data)
python-multipart>=0.0.6
